Separate the graph attribute into its own AstNode file
