+++ ee/codegen/src/generators/workflow.ts
@@ -1,10 +1,10 @@
 import { python } from "@fern-api/python-ast";
-import { OperatorType } from "@fern-api/python-ast/OperatorType";
 import { Reference } from "@fern-api/python-ast/Reference";
 import { Type } from "@fern-api/python-ast/Type";
 import { AstNode } from "@fern-api/python-ast/core/AstNode";
 
 import { BasePersistedFile } from "./base-persisted-file";
+import { GraphAttribute } from "./graph-attribute";
 import { WorkflowOutput } from "./workflow-output";
 
 import {
@@ -14,8 +14,6 @@ import {
   PORTS_CLASS_NAME,
 } from "src/constants";
 import { WorkflowContext } from "src/context";
-import { BaseNodeContext } from "src/context/node-context/base";
-import { PortContext } from "src/context/port-context";
 import { BaseState } from "src/generators/base-state";
 import { Inputs } from "src/generators/inputs";
 import { NodeDisplayData } from "src/generators/node-display-data";
@@ -47,7 +45,6 @@ export class Workflow {
   private readonly inputs: Inputs;
   private readonly nodes: WorkflowDataNode[];
   private readonly edgesByPortId: Map<string, WorkflowEdge[]>;
-  private readonly entrypointPortContexts: PortContext[];
   private readonly entrypointNodeEdges: WorkflowEdge[];
   private readonly displayData: WorkflowDisplayData | undefined;
   private readonly entrypointNodeId: string;
@@ -58,14 +55,9 @@ export class Workflow {
     this.displayData = displayData;
 
     const edges = this.workflowContext.workflowRawEdges;
-    const {
-      edgesByPortId,
-      entrypointPortContexts,
-      entrypointNodeEdges,
-      entrypointNodeId,
-    } = this.getEdgesAndEntrypointNodeContexts({ nodes, edges });
+    const { edgesByPortId, entrypointNodeEdges, entrypointNodeId } =
+      this.getEdgesAndEntrypointNodeContexts({ nodes, edges });
     this.edgesByPortId = edgesByPortId;
-    this.entrypointPortContexts = entrypointPortContexts;
     this.entrypointNodeEdges = entrypointNodeEdges;
     this.entrypointNodeId = entrypointNodeId;
   }
@@ -83,7 +75,6 @@ export class Workflow {
       entrypointNodeId,
     ]);
     const edgesByPortId = new Map<string, WorkflowEdge[]>();
-    let entrypointPortContexts: PortContext[] = [];
     const entrypointNodeEdges: WorkflowEdge[] = [];
 
     edges.forEach((edge) => {
@@ -93,17 +84,6 @@ export class Workflow {
       }
 
       if (edge.sourceNodeId === entrypointNodeId) {
-        const targetNodeContext = this.workflowContext.getNodeContext(
-          edge.targetNodeId
-        );
-
-        const newEntrypointPortContexts = Array.from(
-          targetNodeContext.portContextsById.values()
-        );
-
-        entrypointPortContexts = entrypointPortContexts.concat(
-          newEntrypointPortContexts
-        );
         entrypointNodeEdges.push(edge);
       }
 
@@ -116,7 +96,6 @@ export class Workflow {
 
     return {
       edgesByPortId,
-      entrypointPortContexts,
       entrypointNodeEdges,
       entrypointNodeId,
     };
@@ -549,368 +528,19 @@ export class Workflow {
     return Array.from(this.edgesByPortId.values()).flat();
   }
 
-  private buildGraphAttribute(): AstNode | undefined {
-    // This graph attribute generation is a breadth-first traversal of the WorkflowEdges
-    // starting at the entrypoint port edges. We process each one sequentially and incrementally
-    // build up the graph as we go.
-    type GraphEmpty = { type: "empty" };
-    type GraphSet = { type: "set"; values: GraphMutableAst[] };
-    type GraphNodeReference = {
-      type: "node_reference";
-      reference: BaseNodeContext<WorkflowDataNode>;
-    };
-    type GraphPortReference = {
-      type: "port_reference";
-      reference: PortContext;
-    };
-    type GraphRightShift = {
-      type: "right_shift";
-      lhs: GraphMutableAst;
-      rhs: GraphMutableAst;
-    };
-    type GraphMutableAst =
-      | GraphEmpty
-      | GraphSet
-      | GraphNodeReference
-      | GraphPortReference
-      | GraphRightShift;
-
-    let graphMutableAst: GraphMutableAst = { type: "empty" };
-    const edgesQueue = [...this.entrypointNodeEdges];
-    const processedEdges = new Set<WorkflowEdge>();
-
-    while (edgesQueue.length > 0) {
-      const edge = edgesQueue.shift();
-      if (!edge) {
-        continue;
-      }
-
-      const sourceNode =
-        edge.sourceNodeId === this.entrypointNodeId
-          ? null
-          : this.workflowContext.getNodeContext(edge.sourceNodeId);
-
-      const targetNode = this.workflowContext.getNodeContext(edge.targetNodeId);
-      if (!targetNode) {
-        processedEdges.add(edge);
-        continue;
-      }
-
-      const isPlural = (mutableAst: GraphMutableAst): boolean => {
-        return (
-          mutableAst.type === "right_shift" ||
-          (mutableAst.type === "set" && mutableAst.values.every(isPlural))
-        );
-      };
-
-      const getAstSources = (
-        mutableAst: GraphMutableAst
-      ): GraphPortReference[] => {
-        if (mutableAst.type === "empty") {
-          return [];
-        } else if (mutableAst.type === "node_reference") {
-          const defaultPort = mutableAst.reference.defaultPortContext;
-          if (defaultPort) {
-            return [
-              {
-                type: "port_reference",
-                reference: defaultPort,
-              },
-            ];
-          }
-          return [];
-        } else if (mutableAst.type === "set") {
-          return mutableAst.values.flatMap(getAstSources);
-        } else if (mutableAst.type === "right_shift") {
-          return getAstSources(mutableAst.lhs);
-        } else if (mutableAst.type == "port_reference") {
-          return [mutableAst];
-        } else {
-          return [];
-        }
-      };
-
-      const getAstTerminals = (
-        mutableAst: GraphMutableAst
-      ): GraphNodeReference[] => {
-        if (mutableAst.type === "empty") {
-          return [];
-        } else if (mutableAst.type === "node_reference") {
-          return [mutableAst];
-        } else if (mutableAst.type === "set") {
-          return mutableAst.values.flatMap(getAstTerminals);
-        } else if (mutableAst.type === "right_shift") {
-          return getAstTerminals(mutableAst.rhs);
-        } else if (mutableAst.type == "port_reference") {
-          return [
-            {
-              type: "node_reference",
-              reference: mutableAst.reference.nodeContext,
-            },
-          ];
-        } else {
-          return [];
-        }
-      };
-
-      const popSources = (mutableAst: GraphMutableAst): GraphMutableAst => {
-        if (mutableAst.type === "set") {
-          return {
-            type: "set",
-            values: mutableAst.values.map(popSources),
-          };
-        } else if (mutableAst.type === "right_shift") {
-          return mutableAst.rhs;
-        } else {
-          return { type: "empty" };
-        }
-      };
-
-      const popTerminals = (mutableAst: GraphMutableAst): GraphMutableAst => {
-        if (mutableAst.type === "set") {
-          return {
-            type: "set",
-            values: mutableAst.values.map(popTerminals),
-          };
-        } else if (mutableAst.type === "right_shift") {
-          return mutableAst.lhs;
-        } else {
-          return { type: "empty" };
-        }
-      };
-
-      const addEdgeToGraph = (
-        mutableAst: GraphMutableAst,
-        graphSourceNode: BaseNodeContext<WorkflowDataNode> | null
-      ): GraphMutableAst | undefined => {
-        if (mutableAst.type === "empty") {
-          return {
-            type: "node_reference",
-            reference: targetNode,
-          };
-        } else if (mutableAst.type === "node_reference") {
-          if (sourceNode && mutableAst.reference === sourceNode) {
-            const sourceNodePortContext = sourceNode.portContextsById.get(
-              edge.sourceHandleId
-            );
-            if (sourceNodePortContext) {
-              if (sourceNodePortContext.isDefault) {
-                return {
-                  type: "right_shift",
-                  lhs: mutableAst,
-                  rhs: { type: "node_reference", reference: targetNode },
-                };
-              } else {
-                return {
-                  type: "right_shift",
-                  lhs: {
-                    type: "port_reference",
-                    reference: sourceNodePortContext,
-                  },
-                  rhs: { type: "node_reference", reference: targetNode },
-                };
-              }
-            }
-          } else if (sourceNode == graphSourceNode) {
-            return {
-              type: "set",
-              values: [
-                mutableAst,
-                { type: "node_reference", reference: targetNode },
-              ],
-            };
-          }
-        } else if (mutableAst.type === "set") {
-          const newSet = mutableAst.values.map((subAst) => {
-            const newSubAst = addEdgeToGraph(subAst, graphSourceNode);
-            if (!newSubAst) {
-              return { edgeAdded: false, value: subAst };
-            }
-            return { edgeAdded: true, value: newSubAst };
-          });
-          if (newSet.every(({ edgeAdded }) => !edgeAdded)) {
-            if (sourceNode == graphSourceNode) {
-              return {
-                type: "set",
-                values: [
-                  mutableAst,
-                  { type: "node_reference", reference: targetNode },
-                ],
-              };
-            } else {
-              return;
-            }
-          }
-          const newSetAst: GraphSet = {
-            type: "set",
-            values: newSet.map(({ value }) => value),
-          };
-
-          if (isPlural(newSetAst)) {
-            const newAstTerminals = newSetAst.values.flatMap((value) =>
-              getAstTerminals(value)
-            );
-
-            const uniqueAstTerminalIds = new Set(
-              newAstTerminals.map((terminal) => terminal.reference.getNodeId())
-            );
-            if (uniqueAstTerminalIds.size === 1 && newAstTerminals[0]) {
-              // If all the terminals are the same, we can simplify the graph into a
-              // right shift between the set and the target node.
-              return {
-                type: "right_shift",
-                lhs: popTerminals(newSetAst),
-                rhs: newAstTerminals[0],
-              };
-            }
-          }
-
-          return newSetAst;
-        } else if (mutableAst.type === "right_shift") {
-          const newLhs = addEdgeToGraph(mutableAst.lhs, graphSourceNode);
-          if (newLhs) {
-            const newSetAst: GraphSet = {
-              type: "set",
-              values: [mutableAst, newLhs],
-            };
-            if (isPlural(newSetAst)) {
-              const newAstSources = newSetAst.values.flatMap((value) =>
-                getAstSources(value)
-              );
-
-              const uniqueAstSourceIds = new Set(
-                newAstSources.map((source) => source.reference.portId)
-              );
-              if (uniqueAstSourceIds.size === 1 && newAstSources[0]) {
-                // If all the sources are the same, we can simplify the graph into a
-                // right shift between the source node and the set.
-                const portReference = newAstSources[0];
-                return {
-                  type: "right_shift",
-                  lhs: portReference,
-                  rhs: popSources(newSetAst),
-                };
-              }
-            }
-            return newSetAst;
-          } else if (
-            mutableAst.lhs.type == "port_reference" &&
-            sourceNode &&
-            mutableAst.lhs.reference.nodeContext == sourceNode
-          ) {
-            const sourcePortContext = sourceNode.portContextsById.get(
-              edge.sourceHandleId
-            );
-            if (sourcePortContext) {
-              return {
-                type: "set",
-                values: [
-                  mutableAst,
-                  {
-                    type: "right_shift",
-                    lhs: {
-                      type: "port_reference",
-                      reference: sourcePortContext,
-                    },
-                    rhs: { type: "node_reference", reference: targetNode },
-                  },
-                ],
-              };
-            }
-          } else if (mutableAst.rhs.type === "node_reference") {
-            if (sourceNode && mutableAst.rhs.reference === sourceNode) {
-              return {
-                type: "right_shift",
-                lhs: mutableAst,
-                rhs: { type: "node_reference", reference: targetNode },
-              };
-            }
-          }
-        }
-
-        return;
-      };
-
-      const newMutableAst = addEdgeToGraph(graphMutableAst, null);
-      processedEdges.add(edge);
-
-      if (!newMutableAst) {
-        continue;
-      }
-
-      graphMutableAst = newMutableAst;
-      targetNode.portContextsById.forEach((portContext) => {
-        const edges = this.edgesByPortId.get(portContext.portId);
-        edges?.forEach((edge) => {
-          if (processedEdges.has(edge) || edgesQueue.includes(edge)) {
-            return;
-          }
-          edgesQueue.push(edge);
-        });
-      });
-    }
-
-    const buildAstNode = (mutableAst: GraphMutableAst): AstNode | undefined => {
-      if (mutableAst.type === "empty") {
-        return;
-      }
-
-      if (mutableAst.type === "node_reference") {
-        return python.reference({
-          name: mutableAst.reference.nodeClassName,
-          modulePath: mutableAst.reference.nodeModulePath,
-        });
-      }
-
-      if (mutableAst.type === "port_reference") {
-        return python.reference({
-          name: mutableAst.reference.nodeContext.nodeClassName,
-          modulePath: mutableAst.reference.nodeContext.nodeModulePath,
-          attribute: mutableAst.reference.isDefault
-            ? undefined
-            : [PORTS_CLASS_NAME, mutableAst.reference.portName],
-        });
-      }
-
-      if (mutableAst.type === "set") {
-        return python.TypeInstantiation.set(
-          mutableAst.values.map(buildAstNode).filter(isDefined)
-        );
-      }
-
-      if (mutableAst.type === "right_shift") {
-        const lhs = buildAstNode(mutableAst.lhs);
-        const rhs = buildAstNode(mutableAst.rhs);
-        if (!lhs || !rhs) {
-          return;
-        }
-        return python.operator({
-          operator: OperatorType.RightShift,
-          lhs,
-          rhs,
-        });
-      }
-
-      return;
-    };
-
-    return buildAstNode(graphMutableAst);
-  }
-
   private addGraph(workflowClass: python.Class): void {
     if (this.nodes.length === 0) {
       return;
     }
 
-    const graph = this.buildGraphAttribute();
-
-    if (!graph) {
-      return;
-    }
-
     const graphField = python.field({
       name: "graph",
-      initializer: graph,
+      initializer: new GraphAttribute({
+        edgesByPortId: this.edgesByPortId,
+        entrypointNodeEdges: this.entrypointNodeEdges,
+        entrypointNodeId: this.entrypointNodeId,
+        workflowContext: this.workflowContext,
+      }),
     });
 
     workflowClass.add(graphField);
